phone validation completed

+++ server/src/main/java/server/server/service/verification/UserVerificationService.java
@@ -1,8 +1,11 @@
 package server.server.service.verification;
 
+import com.twilio.Twilio;
+import com.twilio.rest.api.v2010.account.Message;
+import com.twilio.type.PhoneNumber;
+import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Service;
 import server.server.Enum.VerifyStatus;
 import server.server.exception.BadRequestException;
@@ -12,8 +15,8 @@ import server.server.model.UserVerification;
 import server.server.repository.UserVerificationRepository;
 import server.server.service.mail.EmailService;
 import server.server.service.user.UserDetailsService;
-import server.server.service.user.UserService;
 
+import java.util.Random;
 import java.util.UUID;
 
 @Service
@@ -31,6 +34,20 @@ public class UserVerificationService {
 
     @Value("${back-end-url}")
     private String backendUrl;
+
+    @Value("${twilio.account-id}")
+    private String accountId;
+
+    @Value("${twilio.auth-token}")
+    private String authToken;
+
+    @Value("${twilio.phone-no}")
+    private String phoneNo;
+
+    @PostConstruct
+    public void initiateTwilio(){
+        Twilio.init(accountId,authToken);
+    }
 
     public void generateVerificationLink(Long userId){
             String token = UUID.randomUUID().toString();
@@ -59,6 +76,39 @@ public class UserVerificationService {
         Long userId = userVerification.getUserId();
         UserDetails userDetails = userDetailsService.findByUserId(userId);
         userDetails.setEmailVerifyStatus(VerifyStatus.VERIFIED);
+        userDetailsService.saveToDB(userDetails);
+        repository.delete(userVerification);
+    }
+
+    public void generateOTPAndSend(Long userId){
+        UserDetails userDetails = userDetailsService.findByUserId(userId);
+        if(userDetails == null){
+            throw new BadRequestException("User doesn't exist.");
+        }
+        User user = userDetails.getUser();
+        int otp = new Random().nextInt(900000) + 100000 ;
+        UserVerification userVerification = new UserVerification();
+        userVerification.setUserId(userId);
+        userVerification.setOtp(otp);
+        userVerification.setOtpExpiry(System.currentTimeMillis()/1000+10*60);
+        String body = "Your Twitter Verification Code is :"+otp+". expires in 10 mint.";
+        repository.save(userVerification);
+        PhoneNumber to = new PhoneNumber(user.getPoneNumber());
+        PhoneNumber from = new PhoneNumber(phoneNo);
+        Message message = Message.creator(to, from, body).create();
+    }
+
+    public void verifyOtp(int otp){
+        UserVerification userVerification = repository.findByOtp(otp).orElse(null);
+        if(userVerification == null){
+            throw new BadRequestException("Invalid OTP!");
+        }
+        if(userVerification.getOtpExpiry() < System.currentTimeMillis()/1000){
+            throw new BadRequestException("OTP expired, please try again.");
+        }
+        Long userId = userVerification.getUserId();
+        UserDetails userDetails = userDetailsService.findByUserId(userId);
+        userDetails.setPhoneVerifyStatus(VerifyStatus.VERIFIED);
         userDetailsService.saveToDB(userDetails);
         repository.delete(userVerification);
     }
